==> 07-Bayesian-analysis.Rmd <==
# Bayesian Analysis

## Baysian belief

Bayes Theorem deals with the impact of information update on current belief expresses in terms of probability. P(A|B) is the posterior probability of A given knowledge of B. P(A) is the prior probability. P(B|A) is the conditional probability of B given A.
P(A|B)=P(B|A) x P(A)/P(B)

Bayes’ theorem can be expressed in terms of odds and post-test probabilities. This concept had been used in the analysis of the post-test probabilities of stroke related to the use of ABCD2. In this case the likelihood ratio and pretest probability of stroke in patients with TIA is used to calculate the post-test probability.

Table 13.	2 x 2 table
	Disease +	Disease -
Test + eg ABCD2≥4	TP	FP
Test – eg ABCD2<4	FN	TN

The true positive rate (TPR) can be expressed as conditional probability P(T+|D+). TP is the joint distribution of T+ and D+ while TN is the joint distribution of T- and D-.
 
The true negative rate (TNR)  can be expressed as conditional  probability P(T-|D-).
 
In the setting of ABCD2, the pre-test odds is derived from cases of TIA and stroke outcome at 90 days. 
 
The positive likelihood ratio (PLR) in this case is derived from the sensitivity of ABCD2 for stroke and the one minus specificity of ABCD for stroke. One can interpret the likelihood ratio in terms of Bayes theorem.
 
To derive the post-test odds, the PLR is multiplied by the pre-test odds at the voxel level. The post-test odds is given by the product of the pre-test odds and the likelihood ratio. 
 
In turn, the post-test probabilities is calculated from the post-test odds by:  

### Conditional probability

Two events are independent if
 
Or 
 
It follows that the events A and B occur simultaneously P(A∩B) is given by the probability of event A and probability of event B given that A has occurred.
 
If events A and B are independent then P(A∩B) is given by the probability of event A and probability of event B.

#### Bayes Rule

For a sample space (A) containing n different subspaces (A1, A2..An) and A is a subset of the larger sample space B+ and B-, the probability that A is a member of B+an be given by P(A|B+). This can be divised as a tree structure with one branch given by the product of P(B+) and  P(A|B+) and the other P(B-) and P(A|B-). The probability of B is given by the sum of P(B+)P(A|B+) and P(B-)P(A|B).

To make a decision on which of An events to choose, one evaluate the conditional probability of each subset e.g. P(A1∩B)/P(B), P(A2∩B)/P(B)…P(An∩B)/P(B).
The probability of P(B) is given by the sum of P(A1∩B), P(A2∩B)…P(An∩B). Here P(An∩B) is the same as P(An)P(B|An). The subspace with the highest conditional probability may yield the optimal result.
 
#### Conditional independence

The conditional distribution is given by

Conditional independence states that two events (A, B) are independent on a third set (C) if those two sets are independent in their conditional probability distribution given C782.  This is stated as 
 

## Markov model

The Markov chain describes a chain of memoryless states which transit from one state to another without dependency on previous states. The Markov property The transition matrix describes the probabilities of changing from one state to another. A property of this Markov matrix is that the column data sum to one. An example is provided here $\left[\begin{array}{cc}.8 & .7\\.2 &.3\end{array}\right]$. The column probabilities sum to 1. The PageRank method that we  discuss in chapter on Graph Theory is a special form of Markov chain. 

## INLA, Stan and BUGS

In the example here, the GBD 2016 life time risk of stroke is re-used. The model is shown next to the simple linear regression. INLA is used to infer the probability of a set of data given the defined parameters. In the regression example below the linear regression return coefficient of 0.788 and the INLA version returns a mean value of 0.789 as well. Note that Bayesian methods do not provide p-values.

### Linear regression

```{r 07-Bayesian-analysis-1}
library(INLA)

load("./Data-Use/world_stroke.Rda")

#perform ordinary linear regression for comparison
fit<-lm(LifeExpectancy~MeanLifetimeRisk, data =world_sfdf)
summary(fit)
```
Here the output of Bayesian analysis contrasts with the frequentist output above.

```{r 07-Bayesian-analysis-1-1}

#need to subset data as world_sfdf is of class "sf"         "tbl_df"     "tbl" "data.frame"
fitINLA<-inla(LifeExpectancy~MeanLifetimeRisk, family = "gaussian", data =world_sfdf[,c(23,12)])

summary(fitINLA)
```
Performing Bayesian analysis with _rstan_ can be tricky as the text file for the stan code require an empty line at the end. The _brms_ library contains the same precompiled _stan_ code and is easier to run.

```{r 07-Bayesian-analysis-1-2, eval=F}
#This may not run if there is incompatibility between Rtools and version of R
#check  system("g++ -v")
library(brms)

fitBRM<-brm(LifeExpectancy~MeanLifetimeRisk,  data =world_sfdf[,c(23,12)])

summary(fitBRM)
```

Extract the posterior samples for plotting.

```{r 07-Bayesian-analysis-1-3, include=F, eval=F}
library(tidyverse)
post_samples_BRM <- brms::posterior_samples(fitBRM) 

post_samples_BRM %>% 
  select(-lp__) %>% 
  pivot_longer(cols = everything()) %>% 
  ggplot(aes(x = value)) +
  geom_density() +
  facet_wrap(~name, scales = "free")
```

### Logistic regression

This example illustrates the use of INLA for logistic regression.

```{r 07-Bayesian-analysis-2, warning=F}

#library(INLA)
data("BreastCancer",package = "mlbench")
colnames(BreastCancer)

#note Class is benign or malignant of class factor
#need to convert this to numeric values


#first convert to character
BreastCancer$Class<-as.character(BreastCancer$Class)

BreastCancer$Class[BreastCancer$Class=="benign"]<-0
BreastCancer$Class[BreastCancer$Class=="malignant"]<-1



#convert factors to numeric
BreastCancer2<-lapply(BreastCancer[,-c(1,7)], as.numeric)
BreastCancer2<-as.data.frame(BreastCancer2)

#return Class to data
#convert character back to numeric
BreastCancer2$Class<-as.numeric(BreastCancer$Class)

Dx<-inla(Class ~Epith.c.size+Cl.thickness+Cell.shape, family="binomial", 
         data = BreastCancer2)

summary(Dx)
```

The same data is used for logistic regression with stan.

```{r 07-Bayesian-analysis-2-1, warning=F, eval=F}
DxBRM<-brm(Class ~Epith.c.size+Cl.thickness+Cell.shape, 
           family=bernoulli(link = "logit"), data = BreastCancer2)

summary(DxBRM)
```


Extract the posterior sample from logistic regression.

```{r  07-Bayesian-analysis-2-2, warning=F, eval=F}
post_samples_DxBRM <- brms::posterior_samples(DxBRM)

post_samples_DxBRM %>% 
  select(-lp__) %>% 
  pivot_longer(cols = everything()) %>% 
  ggplot(aes(x = value)) +
  geom_density() +
  facet_wrap(~name, scales = "free")

```

### Mixed model

Plotting world_sfdf to identify characteristics of the data

```{r 07-Bayesian-analysis-3, warning=F}
library(ggplot2)
ggplot(data=world_sfdf, aes(x=LifeExpectancy, y=MeanLifetimeRisk,
              color=Income, shape=Continent))+                                               
  geom_point()+
  geom_jitter()

```


Intercept model with INLA

```{r 07-Bayesian-analysis-3-1}

# Set prior on precision
prec.prior <- list(prec = list(param = c(0.001, 0.001)))

Inla_Income<-inla(MeanLifetimeRisk~1+LifeExpectancy+f(Income,  
    model = "iid",
    hyper = prec.prior), 
    data =world_sfdf[,c(23,12, 15, 20)], 
    control.predictor = list(compute = TRUE))

summary(Inla_Income)
  
```

Intercept model with BRMS

```{r 07-Bayesian-analysis-3-2, eval=F}

BRMS_Income <- brm(MeanLifetimeRisk ~ 1 + LifeExpectancy+(1 | Income), 
                              data   = world_sfdf, 
                              warmup = 100, 
                              iter   = 200, 
                              chains = 2, 
                              inits  = "random",
                              cores  = 2)  #use 2 CPU cores simultaneously instead of just 1.

summary(BRMS_Income)
```

```{r  07-Bayesian-analysis-3-3, warning=F, eval=F}
post_samples_BRMS_Income <- brms::posterior_samples(BRMS_Income)

post_samples_BRMS_Income %>% 
  select(-lp__) %>% 
  pivot_longer(cols = everything()) %>% 
  ggplot(aes(x = value)) +
  geom_density() +
  facet_wrap(~name, scales = "free")

```

### Bayesian Metaanalysis 

A Bayesian approach towards metaanalysis is provided below using the package _meta4diag_ [@2015arXiv151206220G]. This approach uses the Integrated Nested Laplacian Approximations (INLA). This package takes a has an advantage over the _mada_ package which does not provide a bivariate method for performing summary sensitivity and specificity. 

```{r 07-Bayesian-analysis-4}
#install.packages("INLA",repos=c(getOption("repos"),INLA="https://inla.r-inla-download.org/R/stable"), dep=TRUE)
library(meta4diag)
library(INLA)

#data from spot sign metaanalysis
Dat<-read.csv("./Data-Use/ss150718.csv")

#remove duplicates
dat<-subset(Dat, Dat$retain=="yes") 

#the data can be viewed under res$data
res <- meta4diag(data = dat) 

#perform SROC
SROC(res, crShow = T)
```

Forest plot of sensitivity using the _meta4diag_ package.

```{r 07-Bayesian-analysis-4-1}
#sensitivity is specified under accuracy.type
#note that there are several different forest plot: mada, metafor, meta4diag
meta4diag::forest(res, accuracy.type="sens", est.type="mean", p.cex="scaled", p.pch=15, p.col="black",
    nameShow="right", dataShow="center", estShow="left", text.cex=1,
    shade.col="gray", arrow.col="black", arrow.lty=1, arrow.lwd=1,
    cut=TRUE, intervals=c(0.025,0.975),
    main="Forest plot of Sensitivity", main.cex=1.5, axis.cex=1)
```

Forest plot of specificity using the _meta4diag_ package.

```{r 07-Bayesian-analysis-4-2}
#specificity is specified under accuracy.type
#note that there are several different forest plot: mada, metafor, meta4diag
meta4diag::forest(res, accuracy.type="spec", est.type="mean", p.cex="scaled", p.pch=15, p.col="black",
    nameShow="right", dataShow="center", estShow="left", text.cex=1,
    shade.col="gray", arrow.col="black", arrow.lty=1, arrow.lwd=1,
    cut=TRUE, intervals=c(0.025,0.975),
    main="Forest plot of Specificity", main.cex=1.5, axis.cex=1)
```

### Cost

The example below is provided by _hesim_.

```{r 07-Bayesian-analysis-5}
library("hesim")
library("data.table")
strategies <- data.table(strategy_id = c(1, 2))
n_patients <- 1000
patients <- data.table(patient_id = 1:n_patients,
          age = rnorm(n_patients, mean = 70, sd = 10),
          female = rbinom(n_patients, size = 1, prob = .4))
states <- data.table(state_id = c(1, 2),
                     state_name = c("Healthy", "Sick")) 
# Non-death health states
tmat <- rbind(c(NA, 1, 2),
              c(3, NA, 4),
              c(NA, NA, NA))
colnames(tmat) <- rownames(tmat) <- c("Healthy", "Sick", "Dead")
transitions <- create_trans_dt(tmat)
transitions[, trans := factor(transition_id)]
hesim_dat <- hesim_data(strategies = strategies,
                        patients = patients, 
                        states = states,
                        transitions = transitions)
print(hesim_dat)
```
