==> 11-App.Rmd <==
# App

There are several platforms for writing and deploying apps. Rstudio has created a library _Shiny_ to write app. This app can be deployed by creating an account at https://shiny.rstudio.com/.

## Brief introduction to Shiny app

Shiny divides the app to 2 component: _ui_ or user interface web page and _server_ or the engine for running the app. This app is used to assist if a stroke code should be activated. It uses the _reactive_ call within _server_ to switch between activating and deactivating code stroke.

```{r 11-App-1}
library(shiny)
# Define UI for application 
ui <- bootstrapPage(
    titlePanel("ST6-code stroke prototype app"),
    wellPanel(
        helpText("This app helps you decide if you should call a code stroke. it is designed for use at a teaching hospital and may not apply for your specific need. Please check with your local hospital."),
        #stroke      
        textInput('stroke',"Is it a stroke-Please enter Yes or No","No"),
        #walk
        textInput('walk',"Prior to stroke, can the patient walk from one room to another without the help from another person? Walking aids such as four wheelframe and walking stick are allowed? -Please enter Yes or No","Yes"),
        #home
        textInput('nursing',"Is the patient from nursing home-Please enter Yes or No","Yes")
    ),
    #partition screen to 3 columns- sum to 12
    column(9,
           tabPanel(
               textOutput("Hospital"),
               textOutput("Walk"))
    )
  )
# Define server logic 
server <- function(input, output, session) {
    #stroke
    answer<-reactive({
       if(input$stroke=="No") 
          {return ("Your answer is NO to the patient having signs of stroke, please do NOT call code stroke")
     } else {
        return("Please call code stroke")
    }
    })
    output$Hospital<-renderText({return(answer())})
    #walk
    answer1<-reactive({
       if(input$stroke=="No") 
      {return ("Your answer is NO to the patient having signs of stroke, please do NOT call code stroke")
       } else { if(input$walk=="No") {
           return("Your answer is No, indicating that premorbidly the patient is unable to walk unaided, please do NOT call code stroke")
       } else { if(input$nursing=="Yes") {
         return("Your answer is Yes to the patient living in nursing home, please do NOT call code stroke")
       }
         }
        return("Please call code stroke")
        }
    })
    output$Walk<-renderText({return(answer1())})
   }
# Run the application 
shinyApp(ui = ui, server = server)
```

Other available apps can be found at https://gntem3.shinyapps.io/ambmc. It was designed to explore deployment of mobile stroke unit (ambulance equipped with CT scanner for stroke treatment) in Melbourne [@pmid31019489]. The app illustrates the use of slide ruler to enable user interaction. A picture can be inserted into the app by placing the picture inside sub-folder _www_. A web page can be inserted using html coding. Plotly object can be rendered wihtin shiny using the call _renderPlotly_. Leaflet map object can be rendered using the call _renderLeaflet_.
